Delete image15-1024x540.jpg
Hi there, 

you need to delete "market4tell" repository and all the contents immediately! I report abuse of content to github, please delete it before your account is suspended by github. 

The name of this repository belongs to a business and it is against copyright to use contents of the website!!!